move types to root of pages
